Move export functionality to new client
