+++ app/views/Project/Tagging/Export/NewExport/EntriesExportSelection/ExportTypePane/styles.css
@@ -1,0 +1,62 @@
+.export-type-pane {
+    display: flex;
+    flex-direction: column;
+    padding: var(--dui-spacing-medium);
+
+    .export-types-container {
+        display: flex;
+        background-color: var(--dui-color-foreground);
+
+        .content {
+            display: flex;
+
+            .export-type {
+                margin: 0 var(--dui-spacing-small);
+            }
+        }
+    }
+
+    .advanced {
+        display: flex;
+        background-color: var(--dui-color-background);
+
+        .content-settings {
+            background-color: var(--dui-color-foreground);
+
+            .checkbox {
+                padding: var(--dui-spacing-extra-small);
+            }
+        }
+
+        .report-structure {
+            display: flex;
+            background-color: var(--dui-color-foreground);
+
+            .content {
+                display: flex;
+                align-items: flex-start;
+                margin: 0 var(--dui-spacing-small);
+
+                .checkbox {
+                    padding: var(--dui-spacing-extra-small);
+                }
+            }
+        }
+
+        .additional {
+            display: flex;
+            background-color: var(--dui-color-foreground);
+
+            .content {
+                display: flex;
+                flex-direction: column;
+                margin: 0 var(--dui-spacing-small);
+
+                .widget-selection {
+                    display: flex;
+                    flex-wrap: wrap;
+                }
+            }
+        }
+    }
+}
