+++ app/views/Project/Tagging/Export/NewExport/EntriesExportSelection/styles.css
@@ -1,0 +1,65 @@
+.export {
+    display: flex;
+    overflow-y: auto;
+
+    .left {
+        display: flex;
+        flex-basis: 70%;
+        flex-direction: column;
+        overflow-y: auto;
+
+        .left-controls {
+            display: flex;
+            flex-direction: column;
+            overflow-y: auto;
+
+            .section {
+                flex-shrink: 0;
+                margin-bottom: var(--dui-spacing-medium);
+
+                .heading {
+                    text-transform: none;
+
+                    .sub-heading {
+                        margin-left: var(--dui-spacing-medium);
+                        color: var(--dui-color-text-label);
+                    }
+                }
+
+                .content {
+                    display: flex;
+                    justify-content: space-between;
+
+                    .query-input {
+                        display: flex;
+                        flex-basis: 50%;
+                        flex-direction: column;
+                    }
+
+                    .save-and-export {
+                        display: flex;
+                        align-self: flex-end;
+                    }
+                }
+            }
+
+            .leads-table-container {
+                display: flex;
+                flex-direction: column;
+            }
+        }
+
+        .export-button {
+            display: flex;
+            align-self: flex-end;
+            flex-shrink: 0;
+            margin: var(--dui-spacing-medium);
+        }
+    }
+
+    .preview {
+        display: flex;
+        flex-basis: 30%;
+        border-left: var(--dui-width-separator-thin) solid var(--dui-color-separator);
+    }
+}
