+++ app/views/Project/Tagging/Export/styles.css
@@ -1,3 +1,27 @@
-.export {
-    color: var(--dui-color-text);
+.container {
+    display: flex;
+    flex-direction: column;
+
+    .content {
+        display: flex;
+        flex-direction: column;
+
+        .top {
+            display: flex;
+
+            .tab-list {
+                flex-grow: 1;
+            }
+
+            .action-buttons {
+                display: flex;
+                align-self: flex-end;
+                flex-shrink: 0;
+
+                .button {
+                    margin: var(--dui-spacing-small);
+                }
+            }
+        }
+    }
 }
