--- src/newViews/Tagging/Export/NewExport/EntriesExportSelection/ExportTypePane/ExportTypeButton/styles.scss
@@ -1,25 +1,0 @@
-@import '~base-scss/utils';
-
-.export-type-select {
-    display: block;
-    outline: 0;
-    border: 0;
-    background-color: transparent;
-    cursor: pointer;
-    padding: var(--spacing-medium);
-
-    .image {
-        @include transition-medium;
-        width: var(--width-icon-extra-large-alt);
-        height: auto;
-        filter: grayscale(80%);
-    }
-
-    &.active {
-        background-color: var(--color-background-accent-hint);
-
-        .image {
-            filter: grayscale(0%);
-        }
-    }
-}
