--- src/newViews/Tagging/Export/styles.scss
@@ -1,27 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-
-    .content {
-        display: flex;
-        flex-direction: column;
-
-        .top {
-            display: flex;
-
-            .tab-list {
-                flex-grow: 1;
-            }
-
-            .action-buttons {
-                display: flex;
-                align-self: flex-end;
-                flex-shrink: 0;
-
-                .button {
-                    margin: var(--dui-spacing-small);
-                }
-            }
-        }
-    }
-}
